requirements.txt
albumentations==1.3.1
certifi==2024.7.4
charset-normalizer==3.3.2
cycler==0.11.0
fonttools==4.43.0
future==0.18.3
idna==3.6
imageio==2.31.2
iso8601==2.1.0
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.5.3
networkx==2.6.3
numpy==1.22.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.2.0
playsound==1.3.0
progressbar==2.5
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0.1
qudida==0.0.4
requests==2.32.2
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.11.1
serial==0.0.97
six==1.16.0
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==2.2.0
typing-extensions==4.8.0
urllib3==2.0.7
